feat: show waitlist capacity
