+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/EnrollmentCell.tsx
@@ -12,6 +12,8 @@ interface EnrollmentCellProps {
      */
     numOnWaitlist: string;
 
+    numWaitlistCap: string;
+
     /**
      * This is a string because numOnWaitlist is a string. I haven't seen this be "n/a" but it seems possible and I don't want it to break if that happens.
      */
@@ -22,6 +24,7 @@ export const EnrollmentCell = ({
     numCurrentlyEnrolled,
     maxCapacity,
     numOnWaitlist,
+    numWaitlistCap,
     numNewOnlyReserved,
 }: EnrollmentCellProps) => {
     return (
@@ -32,7 +35,11 @@ export const EnrollmentCell = ({
                         {numCurrentlyEnrolled.totalEnrolled} / {maxCapacity}
                     </strong>
                 </Box>
-                {numOnWaitlist !== '' && <Box>WL: {numOnWaitlist}</Box>}
+                {numOnWaitlist !== '' && (
+                    <Box>
+                        WL: {numOnWaitlist} / {numWaitlistCap}
+                    </Box>
+                )}
                 {numNewOnlyReserved !== '' && <Box>NOR: {numNewOnlyReserved}</Box>}
             </Box>
         </TableBodyCellContainer>
